update doc, figures, info computations

+++ update.bat
@@ -1,13 +1,37 @@
-echo 'removing old eschaton'
-echo ' '
+echo ARCH = 'uname -m'
 
-sudo rm -rf /Library/Python/2.7/site-packages/eschaton
+win = "i686"
+mac = "darwin"
 
-echo 'copying updated eschaton to site-packages'
-echo ' '
+echo removing old eschaton
+echo  
 
-cp -R ~/GDrive/eschaton /Library/Python/2.7/site-packages/eschaton
+if [ "$(ARCH)" = 'win' ]; then
+   rmdir -rf C:/python27/lib/site-packages/eschaton
+fi
 
-echo 'installing with setup.py'
+if [ "$(ARCH)" = 'mac' ]; then
+   sudo rm -rf /Library/Python/2.7/site-packages/eschaton
+fi
 
-sudo python /Library/Python/2.7/site-packages/eschaton/setup.py install
+echo copying updated eschaton to site-packages'
+echo  
+
+
+if [ "$(ARCH)" = 'win' ]; then
+   xcopy C:/python27/lib/site-packages/eschaton
+fi
+
+if [ "$(ARCH)" = 'mac' ]; then
+   cp -R ~/GDrive/eschaton /Library/Python/2.7/site-packages/eschaton
+fi
+
+echo installing with setup.py
+
+if [ "$(ARCH)" = 'win' ]; then
+   python /Library/Python/2.7/site-packages/eschaton/setup.py install
+fi
+
+if [ "$(ARCH)" = 'mac' ]; then
+   sudo python /Library/Python/2.7/site-packages/eschaton/setup.py install
+fi
